SystemCode/WingSpan-Django-Web-Application/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
branca==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.3
djangorestframework==3.15.0
et-xmlfile==1.1.0
executing==2.0.1
folium==0.16.0
fonttools==4.50.0
git-lfs==1.6
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
immutabledict==4.2.0
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.6.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.4
openai==1.14.1
openpyxl==3.1.2
ortools==9.9.3963
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
plotly==5.20.0
prompt-toolkit==3.0.43
protobuf==5.26.0
psutil==5.9.7
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
requests==2.31.0
scikit-learn==1.4.1.post1
scikit-surprise==1.1.3
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.4.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.12
whitenoise==6.6.0
xyzservices==2024.4.0
